requirements.txt
# requirements.txt
# --- Основной фреймворк и сервер ---
fastapi==0.135.1.0
uvicorn[standard]==0.42  # ASGI сервер, с доп. зависимостями для скорости

# --- Базы данных и драйверы ---
asyncpg==0.31.0            # Асинхронный драйвер для PostgreSQL
sqlalchemy==2.0.48       # Основная ORM
psycopg2-binary==2.9.11  # PostgreSQL драйвер для Python
alembic==1.18.4          # Для миграций БД
redis==7.3.0             # Клиент для Redis
aiosqlite==0.22.1        # Для тестов БД

# --- Аутентификация и безопасность ---
python-jose[cryptography]==3.5.0  # Для работы с JWT (используем cryptography внутри)
pwdlib[argon2, bcrypt]==0.3.0      # Хеширование паролей (будем использовать Argon2 через нее)
cryptography==46.0.5        # Для E2EE шифрования (самая свежая на момент 2026)
python-multipart==0.0.22    # Для обработки форм (нужно для загрузки файлов)

# --- Фоновые задачи и уведомления ---
httpx==0.28.1               # Для HTTP запросов (например, к Firebase)
firebase-admin==7.2.0       # Для работы с Firebase (Push-уведомления)

# --- Валидация данных (Pydantic) ---
pydantic==2.12.5
pydantic-settings==2.13.1    # Для удобной работы с конфигами


# --- Celery для фоновых задач ---
celery==5.6.2   # Основная библиотека Celery
prometheus-client==0.24.1 # Для метрик
circuitbreaker==2.1.3
opentelemetry-sdk==1.40.0
opentelemetry-instrumentation-celery==0.61b0
opentelemetry-instrumentation-openai
celery-types == 0.26.0

# --- Логирование и мониторинг (пока минимум) ---
loguru==0.7.3               # Красивые логи (опционально, но удобно)

# --- Облачное/S3 хранилище ---
minio==7.2.20              # MinIO Python SDK для работы с S3-совместимыми хранилищами (аватары, медиафайлы)

# --- Для тестов ---
pytest==9.0.2
pytest-asyncio==1.3.0     # Чтобы тестировать асинхронный код
pytest-cov==7.0.0           # Для оценки покрытия тестами
opentelemetry-instrumentation-openai

# Фалы
pillow==12.1.1